Add a USE_LOCAL_DB flag to back-end for local debugging
Add the line 'USE_LOCAL_DB=true' to file 'api/.env' to enable this
feature. If the variable is not present in that file, the server behaves
as if it is set to 'false'.

This allows developers to test their code however they want without
polluting our MongoDB Cluster, which hopefully will encourage them to do
more testing before committing their changes.

This would also allow us to run unit tests in the CI without
- adding Travis CI's IP address to the Cluster
- adding credentials to the Cluster to the CI